Crate libp2p_mdns

mDNS is a protocol defined by RFC 6762 that allows querying nodes that correspond to a certain domain name.

In the context of libp2p, the mDNS protocol is used to discover other nodes on the local network that support libp2p.

Usage

This crate provides the Mdns struct which implements the NetworkBehaviour trait. This struct will automatically discover other libp2p nodes on the local network.

Re-exports

pub use self::service::MdnsService;

Modules

service

Structs

Mdns
A NetworkBehaviour for mDNS. Automatically discovers peers on the local network and adds them to the topology.

Enums

MdnsEvent
Event that can be produced by the Mdns behaviour.