Struct libp2p_mdns::Mdns

pub struct Mdns<TSubstream> { /* fields omitted */ }

A NetworkBehaviour for mDNS. Automatically discovers peers on the local network and adds them to the topology.

Methods

impl<TSubstream> Mdns<TSubstream>

pub fn new() -> Result<Mdns<TSubstream>>

Builds a new Mdns behaviour.

pub fn has_node(&self, peer_id: &PeerId) -> bool

Returns true if the given PeerId is in the list of nodes discovered through mDNS.