Crate libp2p_mdns

Implementation of the libp2p-specific mDNS protocol.

mDNS is a protocol defined by RFC 6762 that allows querying nodes that correspond to a certain domain name.

In the context of libp2p, the mDNS protocol is used to discover other nodes on the local network that support libp2p.

Usage

This crate provides a Mdns and TokioMdns, depending on the enabled features, which implements the NetworkBehaviour trait. This struct will automatically discover other libp2p nodes on the local network.

Modules

tokio

The type of a Behaviour using the tokio implementation.

Structs

Behaviour

A NetworkBehaviour for mDNS. Automatically discovers peers on the local network and adds them to the topology.

Config

Configuration for mDNS.

Enums

Event

Event that can be produced by the Mdns behaviour.

Constants

IPV4_MDNS_MULTICAST_ADDRESS

IPV6_MDNS_MULTICAST_ADDRESS